Enhance thread management and debugging capabilities
- Improved the handling of thread states during scheduling and added checks for invalid states to prevent orphaned threads from causing issues.
- Updated thread exit handling in `proc_thread_exit` to prevent thread0 from exiting while other threads are still active, and added tracing for cleanup operations.
- Introduced functions to check for orphaned threads and handle their states appropriately, ensuring that threads are correctly added to the ready queue if they are runnable but not present.
- Added detailed tracing for thread state changes in `add_to_ready_queue`, `remove_from_ready_queue`, and `is_in_ready_queue` functions to improve debugging.
- Removed unused condition variable functions and related structures to streamline synchronization code.
- Commented out some TRACE_THREAD calls in syscall functions to reduce log verbosity.

+++ sys/mint/proc.h
@@ -117,8 +117,8 @@ struct thread {
     /* Thread join fields */
     void *retval;                /* Return value from proc_thread_exit */
     struct thread *joiner;       /* Thread that is joining this thread */
-    int detached;                /* Whether thread is detached */
-    int joined;                  /* Whether thread has been joined */
+    short detached;                /* Whether thread is detached */
+    short joined;                  /* Whether thread has been joined */
     void **join_retval;          /* Where to store return value for joiner */
     
     /* Thread-specific data */
